[TRTLLM-12533][refactor] Move rgba_to_rgb/convert_image_mode to media_io to eliminate circular import with utils
