+++ tensorrt_llm/grpc/grpc_servicer.py
@@ -28,7 +28,7 @@ from typing import List, Union
 import grpc
 
 from tensorrt_llm.executor.result import Logprob, TokenLogprobs
-from tensorrt_llm.inputs.utils import _load_and_convert_image
+from tensorrt_llm.inputs.media_io import _load_and_convert_image
 from tensorrt_llm.logger import logger
 
 from . import trtllm_service_pb2, trtllm_service_pb2_grpc
